+++ src/py_openocd_client/__init__.py
@@ -6,6 +6,7 @@ from .errors import (  # noqa: F401
     OcdCommandFailedError,
     OcdCommandTimeoutError,
     OcdConnectionError,
+    OcdEmptyResponseError,
     OcdInvalidResponseError,
 )
 from .types import BpInfo, BpType, OcdCommandResult, WpInfo, WpType  # noqa: F401