+++ tests_unit/test_errors.py
@@ -1,6 +1,12 @@
 # SPDX-License-Identifier: MIT
 
-from py_openocd_client import OcdCommandFailedError, OcdCommandResult
+import pytest
+
+from py_openocd_client import (
+    OcdCommandFailedError,
+    OcdCommandResult,
+    OcdInvalidResponseError,
+)
 
 
 def test_ocd_command_error_to_string():
@@ -10,3 +16,13 @@ def test_ocd_command_error_to_string():
     cmd_error = OcdCommandFailedError(cmd_result)
 
     assert str(cmd_error) == "OpenOCD command failed: 'my_cmd' (error code: 8)"
+
+
+def test_invalid_response_deprecated_property_out():
+    exc = OcdInvalidResponseError("Exception message", "raw_cmd", "Raw output")
+
+    # "out" is deprecated in favor of "raw_out"
+    with pytest.warns(DeprecationWarning):
+        out = exc.out
+    raw_out = exc.raw_out
+    assert raw_out == out